Special:InactiveUsers: respect unchecked "Hide authenticated"
Due to the way HTTP works, it distinguisih whether boolean values are
false because they were not set at all, or whether they are false
because they have been unchecked.

Work around this by inferring whether this form is shown the first time,
and default "Hide authenticated users" to false from the second time on
(so that the user's choice will be respected even if the checkbox is
unchecked).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>