Update extension submodules
I am seeing various bugs with the pinned commits using deprecated method
calls, etc. Updating to the latest master branch of each brings in code
which does not suffer from these deprecated method usages. So it appears
that the MediaWiki developers are not updating the submodule pointers of
the main mediawiki repository, even though the issues are being
corrected in each respective extension's source repository.

The solution is simple, though: update all the submodules to the latest.
(All the submodules we use, anyway.)