Remove EmbedVideo extension
The Widgets extension can do (almost) everything it can do,
in a much more flexible way. The wiki has been updated to use
{{#widget:YouTube|id=...}} in favor of {{#ev:youtube|...}} everywhere.

It is also nice to eliminate EmbedVideo because it exists in a third
party repository (outside the wikimedia org), and does not seem to be
actively maintained: I had to hack it to make it compatible with the
latest MediaWiki, due to usage of now-removed wfMsg* functions.