Update which skins are available
The MonoBook and Nostalgia skins are removed.
There is no chance we will use them, and IMHO they are pretty ugly.

The erudite skin is added; it might be a good candidate for default
skin. Aside from looking good on mobile devices, it also supports a
top menu bar, although I couldn't immediately figure it out.